fix(pipeline): isolate shared arc limits for browser use (#6769)
