+++ server/lib/storyArcLimits.js
@@ -1,0 +1,17 @@
+/** Canonical length and count caps for story arcs, seasons, and episodes. */
+export const ARC_LIMITS = Object.freeze({
+  LOGLINE_MAX: 500,
+  SUMMARY_MAX: 8000,
+  PROTAGONIST_ARC_MAX: 4000,
+  THEME_MAX: 100,
+  THEMES_PER_ARC_MAX: 20,
+  SEASON_TITLE_MAX: 200,
+  SEASON_LOGLINE_MAX: 500,
+  SEASON_SYNOPSIS_MAX: 8000,
+  SEASON_ENDING_HOOK_MAX: 1000,
+  SEASON_NUMBER_MAX: 99,
+  SEASON_EPISODE_COUNT_MAX: 999,
+  SEASONS_PER_SERIES_MAX: 50,
+  EPISODE_LOGLINE_MAX: 500,
+  EPISODE_SYNOPSIS_MAX: 4000,
+});
